Refactor lib.rs
