+++ polyglot/piranha/src/models/rule_store.rs
@@ -11,11 +11,14 @@ Copyright (c) 2022 Uber Technologies, Inc.
  limitations under the License.
 */
 
-use std::{collections::HashMap, path::Path};
+use std::{collections::HashMap, path::{Path, PathBuf}};
 
 use colored::Colorize;
 use getset::Getters;
+use itertools::Itertools;
+use jwalk::WalkDir;
 use log::{debug, info, trace};
+use regex::Regex;
 use tree_sitter::{Language, Query};
 
 use crate::{
@@ -25,7 +28,7 @@ use crate::{
     rule_graph::RuleGraph,
     scopes::{ScopeGenerator, ScopeQueryGenerator},
   },
-  utilities::{read_toml, MapOfVec},
+  utilities::{read_toml, MapOfVec, read_file},
 };
 
 use super::{
@@ -183,6 +186,86 @@ impl RuleStore {
       .map(|(a, b)| (a.to_string(), b.to_string()))
       .collect();
     let _ = &self.global_tags.extend(global_substitutions);
+  }
+
+  /// To create the current set of global rules, certain substitutions were applied.
+  /// This method creates a regex pattern matching these substituted values.
+  ///
+  /// At the directory level, we would always look to perform global rules. However this is expensive because
+  /// it requires parsing each file. To overcome this, we apply this simple
+  /// heuristic to find the (upper bound) files that would match one of our current global rules.
+  /// This heuristic reduces the number of files to parse.
+  ///
+  pub(crate) fn get_grep_heuristics(&self) -> Regex {
+    let reg_x = self
+      .global_rules()
+      .iter()
+      .flat_map(|r| r.grep_heuristics())
+      .sorted()
+      //Remove duplicates
+      .dedup()
+      //FIXME: Dirty trick to remove true and false. Ideally, grep heuristic could be a field in itself for a rule.
+      // Since not all "holes" could be used as grep heuristic.
+      .filter(|x| {
+        !x.is_empty() && !x.to_lowercase().eq("true") && !x.to_lowercase().as_str().eq("false")
+      })
+      .join("|");
+    Regex::new(reg_x.as_str()).unwrap()
+  }
+
+  pub(crate) fn does_file_extension_match(&self, de: &jwalk::DirEntry<((), ())>) -> bool {
+    de.path()
+      .extension()
+      .and_then(|e| {
+        e.to_str()
+          .filter(|x| x.eq(&self.piranha_args().get_language()))
+      })
+      .is_some()
+  }
+
+  /// Checks if any global rule has a hole
+  pub(crate) fn any_global_rules_has_holes(&self) -> bool {
+    self.global_rules().iter().any(|x| !x.holes().is_empty())
+  }
+
+  /// Gets all the files from the code base that (i) have the language appropriate file extension, and (ii) contains the grep pattern.
+  /// Note that `WalkDir` traverses the directory with parallelism.
+  /// If all the global rules have no holes (i.e. we will have no grep patterns), we will try to find a match for each global rule in every file in the target.
+  pub(crate) fn get_relevant_files(&self, path_to_codebase: &str) -> HashMap<PathBuf, String> {
+    let _path_to_codebase = Path::new(path_to_codebase).to_path_buf();
+
+    //If the path_to_codebase is a file, then execute piranha on it 
+    if _path_to_codebase.is_file() {
+      return HashMap::from_iter([(
+        _path_to_codebase.clone(),
+        read_file(&_path_to_codebase).unwrap()
+      )]);
+    }
+    let mut files: HashMap<PathBuf, String> = WalkDir::new(path_to_codebase)
+      // Walk over the entire code base
+      .into_iter()
+      // Ignore errors
+      .filter_map(|e| e.ok())
+      // Filter files with the desired extension
+      .filter(|de| self.does_file_extension_match(de))
+      // Read the file
+      .map(|f| (f.path(), read_file(&f.path()).unwrap()))
+      .collect();
+
+    if self.any_global_rules_has_holes() {
+      let pattern = self.get_grep_heuristics();
+      files = files
+        .iter()
+        // Filter the files containing the desired regex pattern
+        .filter(|x| pattern.is_match(x.1.as_str()))
+        .map(|(x, y)| (x.clone(), y.clone()))
+        .collect();
+    }
+    debug!(
+      "{}",
+      format!("{} files will be analyzed.", files.len()).green()
+    );
+    files
   }
 }
